# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ANAM-BF - API de la plateforme agrométéo",
    "description": "\nAPI de la plateforme agrométéorologique de l'ANAM-BF.\n\n**Conventions**\n- Format JSON, UTF-8. Dates en UTC au format ISO 8601.\n- Authentification : en-tête `Authorization: Bearer <access_token>` (voir la section *Authentification*).\n  Les contenus publiés (bulletins, alertes, avis, carte, zones) sont **publics** : consultables sans compte.\n- Erreurs : `{\"detail\": \"message lisible\"}` avec le code HTTP adapté (401 non connecté, 403 droits insuffisants,\n  404 introuvable, 409 conflit d'état, 422 données invalides, 429 trop de tentatives).\n- Listes paginées : paramètres `limit` et `offset`, réponse `{\"total\": N, \"items\": [...]}`.\n- Langues : `fr` (français), `en` (anglais), `mos` (mooré).\n- Génération audio/vidéo : asynchrone. Lancer, puis relire le contenu jusqu'à `media.status` = `ready`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/zones": {
      "get": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "List Zones",
        "description": "Zones (communes pilotes et régions) utilisables pour les alertes et la carte. Public.\n\n**Accès :** Public.",
        "operationId": "list_zones_zones_get",
        "parameters": [
          {
            "name": "pilot_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Seulement les 5 communes pilotes",
              "default": false,
              "title": "Pilot Only"
            },
            "description": "Seulement les 5 communes pilotes"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ZoneOut"
                  },
                  "title": "Response List Zones Zones Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "Create Zone",
        "description": "**Accès :** Connexion requise, permission `config:manage` (rôles : Administrateur).",
        "operationId": "create_zone_zones_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZoneWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/zones/{zone_id}": {
      "get": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "Get Zone",
        "description": "Détail d'une zone, avec son contour GeoJSON s'il est renseigné. Public.\n\n**Accès :** Public.",
        "operationId": "get_zone_zones__zone_id__get",
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Zone Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "Update Zone",
        "description": "Modifie une zone (ex. renseigner ses coordonnées ou son contour GeoJSON fournis par l'ANAM).\n\n**Accès :** Connexion requise, permission `config:manage` (rôles : Administrateur).",
        "operationId": "update_zone_zones__zone_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Zone Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ZoneUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "Delete Zone",
        "description": "**Accès :** Connexion requise, permission `config:manage` (rôles : Administrateur).",
        "operationId": "delete_zone_zones__zone_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "zone_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Zone Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/alert-types": {
      "get": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "List Alert Types",
        "description": "Types d'alerte (orages, fortes pluies, inondations, vents violents, poussière, chaleur, sécheresse…). Public.\n\n**Accès :** Public.",
        "operationId": "list_alert_types_alert_types_get",
        "parameters": [
          {
            "name": "include_inactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inactive"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AlertTypeOut"
                  },
                  "title": "Response List Alert Types Alert Types Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "Create Alert Type",
        "description": "**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "create_alert_type_alert_types_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertTypeWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertTypeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alert-types/{type_id}": {
      "patch": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "Update Alert Type",
        "description": "Modifie les libellés d'un type ; `active=false` le retire des choix sans casser les alertes existantes.\n\n**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "update_alert_type_alert_types__type_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Type Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertTypeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertTypeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prevention-messages": {
      "get": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "List Prevention",
        "description": "Messages de prévention par type d'alerte. Public.\n\n**Accès :** Public.",
        "operationId": "list_prevention_prevention_messages_get",
        "parameters": [
          {
            "name": "alert_type_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Alert Type Id"
            }
          },
          {
            "name": "alert_type_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Alert Type Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreventionOut"
                  },
                  "title": "Response List Prevention Prevention Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "Create Prevention",
        "description": "**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "create_prevention_prevention_messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreventionWrite"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreventionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prevention-messages/{message_id}": {
      "patch": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "Update Prevention",
        "description": "**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "update_prevention_prevention_messages__message_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreventionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreventionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "Delete Prevention",
        "description": "**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "delete_prevention_prevention_messages__message_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/prevention-messages/{message_id}/translate": {
      "post": {
        "tags": [
          "Zones, carte et référentiels"
        ],
        "summary": "Translate Prevention",
        "description": "Propose une traduction automatique (anglais : MyMemory, mooré : CITADEL/NLLB) à partir du texte français.\nRésultat à relire avant diffusion. Les traductions déjà saisies ne sont pas écrasées sauf `overwrite=true`.\n\n**Accès :** Connexion requise, permission `content:manage` (rôles : Agent ANAM, Administrateur).",
        "operationId": "translate_prevention_prevention_messages__message_id__translate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Message Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreventionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ZoneDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/ZoneKind"
          },
          "is_pilot": {
            "type": "boolean",
            "title": "Is Pilot"
          },
          "commune_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Commune Names"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "is_pilot",
          "commune_names",
          "latitude",
          "longitude",
          "geometry"
        ],
        "title": "ZoneDetail"
      },
      "ZoneKind": {
        "type": "string",
        "enum": [
          "commune",
          "region"
        ],
        "title": "ZoneKind"
      },
      "ZoneOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/ZoneKind"
          },
          "is_pilot": {
            "type": "boolean",
            "title": "Is Pilot"
          },
          "commune_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Commune Names"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "is_pilot",
          "commune_names",
          "latitude",
          "longitude"
        ],
        "title": "ZoneOut"
      },
      "ZoneUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ZoneKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_pilot": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Pilot"
          },
          "commune_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commune Names"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          }
        },
        "type": "object",
        "title": "ZoneUpdate"
      },
      "ZoneWrite": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Name"
          },
          "kind": {
            "$ref": "#/components/schemas/ZoneKind",
            "default": "commune"
          },
          "is_pilot": {
            "type": "boolean",
            "title": "Is Pilot",
            "default": false
          },
          "commune_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Commune Names",
            "default": []
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "geometry": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry",
            "description": "GeoJSON Polygon ou MultiPolygon"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ZoneWrite"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}